requirements.txt
matplotlib >= 3.2.1
torchvision >= 0.7.0+cu101
pandas >= 1.1.3
torchsummary >= 1.5.1
opencv-python >= 3.4.11.45
numpy >= 1.19.5
mediapipe >= 0.8.0
Pillow >= 7.2.1
scikit-learn >= 0.23.2
torch >= 1.6.0
